api: add object.go
